+++ config/silverstripe-5-3.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-use Netwerkstatt\SilverstripeRector\Rector\Misc\RenameAddFieldsToTabWithoutArrayParamRector;
+use Netwerkstatt\SilverstripeRector\Rector\Misc\RenameFieldListMethodsWithoutArrayParamRector;
 use Rector\Config\RectorConfig;
 use Rector\Renaming\Rector\Name\RenameClassRector;
 
 return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->rule(RenameAddFieldsToTabWithoutArrayParamRector::class);
+    $rectorConfig->rule(RenameFieldListMethodsWithoutArrayParamRector::class);
 
     $rectorConfig->ruleWithConfiguration(RenameClassRector::class, [
         'SilverStripe\ORM\DataExtension' => 'SilverStripe\Core\Extension',